+++ foxes/input/states/__init__.py
@@ -6,7 +6,6 @@ from .single import SingleStateStates
 from .scan import ScanStates
 from .states_table import StatesTable, Timeseries, TabStates
 from .field_data_nc import FieldDataNC
-from .slice_data_nc import SliceDataNC
 from .multi_height import MultiHeightStates, MultiHeightTimeseries
 from .multi_height import MultiHeightNCStates, MultiHeightNCTimeseries
 from .one_point_flow import (